Factor out the runtime options into a new file.
